Crate core_transport

The crate aggregates and composes individual transport level objects and functionality into a unified HoprTransport object with the goal of isolating the transport layer and defining a fully specified transport API.

As such, the transport layer components should be only those that are directly needed in order to:

send and receive a packet, acknowledgement or ticket aggregation request

send and receive a network telemetry request

automate transport level processes

algorithms associated with the transport layer operational management

interface specifications to allow modular behavioral extensions

Re-exports

pub use core_p2p::libp2p;

Modules

adaptors

api

Externally visible API objects containing definitions of the p2p composite HOPR protocol types.

config

Configuration of the HoprTransport.

constants

Constants used and exposed by the crate.

errors

Errors used by the crate.

Structs

ApplicationData

Represents the received decrypted packet carrying the application-layer data.

ChainKeypair

Represents a keypair consisting of a secp256k1 private and public key

ChannelEventEmitter

Event emitter used by the indexer to emit events when an on-chain change on a channel is detected.

HalfKeyChallenge

Represents a challenge for the half-key in Proof of Relay. Half-key challenge is equivalent to a secp256k1 curve point. Therefore, HalfKeyChallenge can be obtained from a HalfKey.

Hash

Represents an Ethereum 256-bit hash value This implementation instantiates the hash via Keccak256 digest.

HoprTransport

Interface into the physical transport mechanism allowing all HOPR related tasks on the transport mechanism, as well as off-chain ticket manipulation.

IndexerActions

Implementor interface for indexer actions

Multiaddr

Representation of a Multiaddr.

Network

The network object storing information about the running observed state of the network, including peers, connection qualities and updates for other parts of the system.

OffchainKeypair

Represents a keypair consisting of an Ed25519 private and public key

OffchainPublicKey

Represents an Ed25519 public key.

PeerStatus

Status of the peer as recorded by the [Network].

PublicNodesResult

TicketStatistics

Ticket statistics data exposed by the ticket mechanism.

Enums

Health

Network health represented with colors, where green is the best and red is the worst possible observed nework quality.

IndexerProcessed

Processed indexer generated events.

IndexerToProcess

Indexer events triggered externally from the crate::HoprTransport object.

NetworkTriggeredEvent

Events generated by the Network object allowing it to physically interact with external systems, including the transport mechanism.

PeerEligibility

PeerOrigin

Actual origin.

TransportOutput

Composite output from the transport layer.

Traits

Keypair

Represents a generic key pair The keypair contains a private key and public key. Must be comparable in constant time and zeroized on drop.

Functions

build_index_updater

Build the index updater mechanism for indexer generated behavior inclusion.

build_network

Build the Network object responsible for tracking and holding the observable state of the physical transport network, peers inside the network and telemetry about network connections.

build_packet_actions

Build the interaction object allowing to process messages.

build_ticket_aggregation

Build the ticket aggregation mechanism and processes.

build_transport_components

decapsulate_p2p_protocol

Remove the p2p/<PeerId> component from a multiaddress

execute_on_tick

Construct an infinitely running background loop producing ticks with a given period with the maximum tick duration at most the period.

p2p_loop

Main p2p loop that instantiates a new libp2p::Swarm instance and sets up listening and reacting pipelines running in a neverending loop future.