Crate core_transport

The crate aggregates and composes individual transport level objects and functionality into a unified crate::HoprTransport object with the goal of isolating the transport layer and defining a fully specified transport API.

As such, the transport layer components should be only those that are directly needed in order to:

send and receive a packet, acknowledgement or ticket aggregation request

send and receive a network telemetry request

automate transport level processes

algorithms associated with the transport layer operational management

interface specifications to allow modular behavioral extensions

Re-exports

pub use crate::helpers::IndexerTransportEvent;

pub use crate::helpers::PeerEligibility;

pub use crate::helpers::TicketStatistics;

pub use hopr_transport_p2p::libp2p;

Modules

config

Configuration of the crate::HoprTransport.

constants

Constants used and exposed by the crate.

errors

Errors used by the crate.

helpers

network_notifier

Structs

AggregatorProxy

ApplicationData

Represents the received decrypted packet carrying the application-layer data.

ChainKeypair

Represents a keypair consisting of a secp256k1 private and public key

HalfKeyChallenge

Represents a challenge for the half-key in Proof of Relay. Half-key challenge is equivalent to a secp256k1 curve point. Therefore, HalfKeyChallenge can be obtained from a HalfKey.

Hash

Represents an Ethereum 256-bit hash value This implementation instantiates the hash via Keccak256 digest.

HoprSwarmWithProcessors

HoprTransport

Interface into the physical transport mechanism allowing all HOPR related tasks on the transport mechanism, as well as off-chain ticket manipulation.

HoprTransportConfig

Multiaddr

Representation of a Multiaddr.

Network

The network object storing information about the running observed state of the network, including peers, connection qualities and updates for other parts of the system.

OffchainKeypair

Represents a keypair consisting of an Ed25519 private and public key

OffchainPublicKey

Represents an Ed25519 public key.

PeerStatus

Status of the peer as recorded by a network component.

Session

SessionClientConfig

Configuration for the session.

SessionId

Unique ID of a specific session.

Enums

Health

Network health represented with colors, where green is the best and red is the worst possible observed nework quality.

HoprTransportProcess

NetworkTriggeredEvent

Events generated by the Network object allowing it to physically interact with external systems, including the transport mechanism.

PathOptions

Send options for the session.

PeerOrigin

Actual origin.

PeerTransportEvent

Processed indexer generated events.

SessionCapability

TransportOutput

Composite output from the transport layer.

TransportSessionError

Enumeration of errors thrown from this library.

Constants

SESSION_USABLE_MTU_SIZE

Traits

Keypair

Represents a generic key pair The keypair contains a private key and public key. Must be comparable in constant time and zeroized on drop.

SendMsg

Functions

execute_on_tick

Construct an infinitely running background loop producing ticks with a given period with the maximum tick duration at most the period.

strip_p2p_protocol

Remove the p2p/<PeerId> component from a multiaddress