Crate core_transport

Re-exports

pub use core_p2p::libp2p;

Modules

adaptors

api

config

constants

errors

Structs

ApplicationData

Represents the received decrypted packet carrying the application-layer data.

ChainKeypair

Represents a keypair consisting of a secp256k1 private and public key

ChannelEventEmitter

This is used by the indexer to emit events when a change on channel entry is detected.

ExternalNetworkInteractions

HalfKeyChallenge

Represents a challenge for the half-key in Proof of Relay. Half-key challenge is equivalent to a secp256k1 curve point. Therefore, HalfKeyChallenge can be obtained from a HalfKey.

Hash

Represents an Ethereum 256-bit hash value This implementation instantiates the hash via Keccak256 digest.

HoprTransport

IndexerActions

Implementor interface for indexer actions

Multiaddr

Representation of a Multiaddr.

Network

OffchainKeypair

Represents a keypair consisting of an Ed25519 private and public key

OffchainPublicKey

Represents an Ed25519 public key.

PeerStatus

PublicNodesResult

TicketStatistics

Enums

Health

IndexerProcessed

IndexerToProcess

NetworkEvent

PeerEligibility

PeerOrigin

TransportOutput

Object representing different types of output from the transport layer

Traits

Keypair

Represents a generic key pair The keypair contains a private key and public key. Must be comparable in constant time and zeroized on drop.

NetworkExternalActions

Functions

build_heartbeat

build_index_updater

build_manual_ping

build_network

build_packet_actions

build_ticket_aggregation

decapsulate_p2p_protocol

execute_on_tick

Represents a periodically timed ticks in a loop with the given period. Could be later extended so it supports multiple different periods and multiple actions.

p2p_loop

Main p2p loop that will instantiate a new libp2p::Swarm instance and setup listening and reacting pipelines running in a neverending loop future.