Function core_transport::p2p_loop

pub async fn p2p_loop(
    version: String,
    me: Keypair,
    network: Arc<RwLock<Network<ExternalNetworkInteractions>>>,
    network_update_input: Receiver<NetworkEvent>,
    indexer_update_input: Receiver<IndexerProcessed>,
    ack_interactions: AcknowledgementInteraction,
    pkt_interactions: PacketInteraction,
    ticket_aggregation_interactions: TicketAggregationInteraction<ResponseChannel<Result<Ticket, String>>, OutboundRequestId>,
    heartbeat_requests: HeartbeatRequester,
    heartbeat_responds: HeartbeatResponder,
    manual_ping_requests: ManualPingRequester,
    manual_ping_responds: HeartbeatResponder,
    my_multiaddresses: Vec<Multiaddr>,
    protocol_cfg: ProtocolConfig,
    on_transport_output: UnboundedSender<TransportOutput>,
    on_acknowledged_ticket: UnboundedSender<AcknowledgedTicket>
)

Main p2p loop that will instantiate a new libp2p::Swarm instance and setup listening and reacting pipelines running in a neverending loop future.

The function represents the entirety of the business logic of the hopr daemon related to core operations.

This future can only be resolved by an unrecoverable error or a panic.